All Codes/pubsub_to_sql/requirements.txt
mysql-connector-python
cloudevents
